refactor(connect-popup): split selectAccount thunks into selectAccountThunks
Move the selectAccount picker thunks (load page, manual account/address,
verify, resolve) out of connectPopupThunks into their own file. This is
behavior-preserving; the shared CONNECT_POPUP_MODULE action prefix moves to
connectPopupTypes so both thunk files can import it.
